server/requirements.txt
ffmpeg-python==0.2.0
pymediainfo==5.1.0

SQLAlchemy==1.4.41
aiosqlite==0.17.0
alembic==1.8.1

pyinstaller==5.4.1
pyinstaller-setuptools==2019.3

pytest==7.1.3
pytest-cov==3.0.0
pytest-alembic==0.8.4
pytest-timeout==2.1.0
pytest-asyncio==0.19.0
pytest-xdist==2.5.0

slack-sdk==3.18.3

python-socketio==5.7.1
aiohttp==3.8.1
mypy==0.971

docopt==0.6.2
requests==2.28.1
